# services/partner-service/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Partner Service API
  version: 0.1.0
  description: |
    Partner and booking management service for Pulcherbook platform.

    This service manages:
    - Partners (companies) registration and information
    - Teams and team members
    - Services offered by partners
    - Company media (logos, images)
    - Appointment booking system
    - Team member availability

    **Note**: Authentication is currently public. Authentication/authorization will be implemented in a future update.
servers:
  - url: http://localhost:3002
    description: Development server
  - url: https://api.pulcherbook.com
    description: Production server
...
paths:
...
  /api/v1/partners/{partner_uid}/shops:
    post:
      summary: Create a new shop
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShopRequest'
      responses:
        '201':
          description: Shop created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '400':
          description: Invalid request

    get:
      summary: Get all shops for a partner
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shop'
  /api/v1/partners/{partner_uid}/shops/{uid}:
    get:
      summary: Get shop by ID
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shop details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '404':
          description: Shop not found

    put:
      summary: Update shop
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopRequest'
      responses:
        '200':
          description: Shop updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '404':
          description: Shop not found

    delete:
      summary: Delete shop
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shop deleted
        '404':
          description: Shop not found
  /api/v1/partners/{partner_uid}/shops/{uid}/enable:
    post:
      summary: Enable shop (set status to ONLINE)
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shop enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '404':
          description: Shop not found
  /api/v1/partners/{partner_uid}/shops/{uid}/disable:
    post:
      summary: Disable shop (set status to OFFLINE)
      tags: [Shops]
      parameters:
        - name: partner_uid
          in: path
          required: true
          schema:
            type: string
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shop disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '404':
          description: Shop not found
components:
  schemas:
...
    Shop:
      type: object
      properties:
        uid:
          type: string
        partner_uid:
          type: string
        shop_name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postal_code:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        instagram:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateShopRequest:
      type: object
      required:
        - shop_name
      properties:
        shop_name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postal_code:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        instagram:
          type: string
    UpdateShopRequest:
      type: object
      properties:
        shop_name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postal_code:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        instagram:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication (TODO - to be implemented)